mlflow_part/requirements.txt
pandas
numpy
scikit-learn
psycopg2-binary
pymongo
mlflow
joblib
 

